fix: changing the project structure to have two packages inside "pkg", one for http and one for server
