+++ .github/workflows/main.yaml
@@ -39,5 +39,7 @@ jobs:
       - uses: actions/setup-go@v5
         with:
           go-version: stable
-      - name: Test with the Go CLI
-        run: go test -v ./pkg
+      - name: Test http server
+        run: go test -v ./pkg/httpserver
+      - name: Test gin server
+        run: go test -v ./pkg/ginserver
